Finish configuring the routing correctly and put the correct naming to the files
