+++ Components/Molecules/NavBar/NavBar.jsx
@@ -15,7 +15,7 @@ export const NavBar = () => {
         </Link>
         <div className='flex gap-6'>
           <button className='text-text px-2 py-1 text-center rounded-t hover:bg-btnhover'>
-            <Link href='/add' className='text-sm'>Agregar integrante</Link>
+            <Link href='/AddMemberScreen' className='text-sm'>Agregar integrante</Link>
           </button>
           <button className='text-text px-2 py-1 text-center rounded-t hover:bg-btnhover'>
             <Link href='/LoginScreen' className='text-sm'>Iniciar sesión</Link>
