+++ app/AddMemberScreen/page.jsx
@@ -1,0 +1,11 @@
+import AddMenber from "@/Components/Screens/General/AddMember/AddMenber";
+import React from "react";
+
+export default function Page () {
+
+    return(
+        <>
+            <AddMenber />
+        </>
+    );
+}
