+++ app/page.jsx
@@ -1,7 +1,5 @@
+import React from "react";
 import LandingPage from "@/Components/Screens/General/LandingPage/LandingPage";
-
-
-
 
 export default function Home() {
   return (
